Order dao and logging
