+++ build.gradle.kts
@@ -23,6 +23,11 @@ dependencies {
 	implementation("org.jetbrains.kotlin:kotlin-reflect")
 	implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
 
+	// kotlin logging as wrapper for slf4j
+	implementation("io.github.microutils:kotlin-logging-jvm:3.0.5")
+	implementation("org.slf4j:slf4j-api:2.0.7")
+	implementation("ch.qos.logback:logback-classic:1.4.14")
+
 	// GraphQl
 	implementation("com.expediagroup:graphql-kotlin-spring-server:7.0.0-alpha.5")
