+++ src/main/kotlin/com/carbonara/core/product/Product.kt
@@ -2,9 +2,6 @@ package com.carbonara.core.product
 
 import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.expediagroup.graphql.generator.scalars.ID
-import org.bson.types.ObjectId
-import org.springframework.data.annotation.Id
-import org.springframework.data.mongodb.core.mapping.Document
 
 @GraphQLDescription("A food box for customers")
 class Product (
@@ -24,23 +21,3 @@ class Product (
     @GraphQLDescription("This product is active right now and can be purchased")
     var isActive: Boolean
 )
-
-@Document("Product")
-data class ProductDto(
-    @Id
-    val productId: ObjectId,
-    val productName: String,
-    val productPrice: Int,
-    val productPictureUrl: String,
-    var isActive: Boolean?
-) {
-    fun toProduct(): Product {
-        return Product(
-            productId = ID(productId.toString()),
-            productName = productName,
-            productPrice = productPrice,
-            productPictureUrl = productPictureUrl,
-            isActive = isActive ?: false
-        )
-    }
-}
